projects/Ip Add Converter Tool/requirements.txt
tkinter
pillow==10.1.0